requirements.txt
fastapi
uvicorn
python-multipart
httpx
pydantic
uuid
python-docx
pymupdf
pandas
langchain
langchain-openai
langchain-community
faiss-cpu
openai
pillow
python-dotenv
python-fasthtml
shad4fast